pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bank-roi-engine"
version = "0.1.0"
description = "Profit-driven customer targeting engine for bank telemarketing campaigns"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.2",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "xgboost>=2.0",
    "imbalanced-learn>=0.12",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "plotly>=5.20",
    "streamlit>=1.35",
    "pyyaml>=6.0",
    "joblib>=1.3",
    "shap>=0.44",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov", "ruff", "mypy"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
